added capybara helper select2 + fix variantautocomplete
