+++ core/app/views/spree/admin/variants/_autocomplete.js.erb
@@ -13,8 +13,8 @@
       <h6 class="variant-name">{{variant.name}}</h6>
 
       <ul class='variant-data'>
-        <li class='variant-sku'><strong>{{t('sku')}}:</strong> {{variant.sku}}</li>
-        <li class='variant-on_hand'><strong>{{t('on_hand')}}:</strong> {{variant.count_on_hand}}</li>
+        <li class='variant-sku'><strong>{{t 'sku'}}:</strong> {{variant.sku}}</li>
+        <li class='variant-on_hand'><strong>{{t 'on_hand' }}:</strong> {{variant.count_on_hand}}</li>
       </ul>
 
       {{#if variant.option_values}}
