+++ core/spec/requests/admin/orders/customer_details_spec.rb
@@ -35,7 +35,7 @@ describe "Customer Details" do
   context "editing an order", :js => true do
     it "should be able to populate customer details for an existing order" do
       click_link "Customer Details"
-      select2("#select-customer", "foobar")
+      select2_search("foobar", :from => "select-customer")
 
       ["ship_address", "bill_address"].each do |address|
         find_field("order_#{address}_attributes_firstname").value.should == "John"
@@ -64,7 +64,7 @@ describe "Customer Details" do
         fill_in "order_#{type}_address_attributes_address2",   :with => "#101"
         fill_in "order_#{type}_address_attributes_city",       :with => "Bethesda"
         fill_in "order_#{type}_address_attributes_zipcode",    :with => "20170"
-        select "Alabama", :from => "order_#{type}_address_attributes_state_id"
+        select2 "Alabama", :from => "order_#{type}_address_attributes_state_id"
         fill_in "order_#{type}_address_attributes_phone",     :with => "123-456-7890"
       end
