+++ core/spec/requests/admin/products/edit/taxons_spec.rb
@@ -22,7 +22,7 @@ describe "Product Taxons" do
 
       find(".select2-search-choice").text.should == taxon_1.name
       selected_taxons.should =~ [taxon_1.id]
-      select2("#product_taxons_field", "Clothing")
+      select2_search("#product_taxons_field", "Clothing")
       click_button "Update"
       selected_taxons.should =~ [taxon_1.id, taxon_2.id]
