+++ core/spec/requests/admin/products/edit/variants_spec.rb
@@ -42,7 +42,7 @@ describe "Product Variants" do
         click_icon :edit
       end
 
-      select2("#product_option_types_field", "color")
+      select2_search("#product_option_types_field", "color")
       click_button "Update"
       page.should have_content("successfully updated!")
