+++ core/spec/spec_helper.rb
@@ -11,6 +11,7 @@ Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
 require 'database_cleaner'
 
 require 'spree/core/testing_support/factories'
+require 'spree/core/testing_support/capybara_ext'
 require 'spree/core/testing_support/controller_requests'
 require 'spree/core/testing_support/authorization_helpers'
 require 'spree/core/testing_support/preferences'
