+++ promo/Gemfile
@@ -1,5 +1,4 @@
 eval(File.read(File.dirname(__FILE__) + '/../common_spree_dependencies.rb'))
-
 gem 'spree_core', :path => '../core'
 
 gemspec
