+++ promo/spec/requests/promotion_adjustments_spec.rb
@@ -74,7 +74,9 @@ describe "Promotion Adjustments" do
     it "should allow an admin to create a single user coupon promo with flat rate discount" do
       fill_in "Name", :with => "Order's total > $30"
       fill_in "Usage Limit", :with => "1"
-      select "Coupon code added", :from => "Event"
+
+      select2 "Coupon code added", :from => "Event Name"
+
       fill_in "Code", :with => "SINGLE_USE"
       click_button "Create"
       page.should have_content("Editing Promotion")
@@ -142,7 +144,8 @@ describe "Promotion Adjustments" do
 
     it "should allow an admin to create an automatic promo with flat percent discount" do
       fill_in "Name", :with => "Order's total > $30"
-      select "Order contents changed", :from => "Event"
+
+      select2 "Order contents changed", :from => "Event Name"
       click_button "Create"
       page.should have_content("Editing Promotion")
 
@@ -232,7 +235,8 @@ describe "Promotion Adjustments" do
 
     it "should allow an admin to create an automatic promo requiring a landing page to be visited" do
       fill_in "Name", :with => "Deal"
-      select "Visit static content page", :from => "Event"
+
+      select2 "Visit static content page", :from => "Event Name"
       fill_in "Path", :with => "content/cvv"
       click_button "Create"
       page.should have_content("Editing Promotion")
@@ -288,12 +292,14 @@ describe "Promotion Adjustments" do
 
     it "should allow an admin to create a promotion that adds a 'free' item to the cart" do
       fill_in "Name", :with => "Bundle"
-      select "Coupon code added", :from => "Event"
+      select2 "Coupon code added", :from => "Event Name"
+
       fill_in "Code", :with => "5ZHED2DH"
       click_button "Create"
       page.should have_content("Editing Promotion")
 
-      select "Create line items", :from => "Add action of type"
+      select2 "Create line items", :from => "Add action of type"
+
       within('#action_fields') { click_button "Add" }
       # Forced narcolepsy, thanks to JavaScript
       sleep(1)
@@ -345,7 +351,7 @@ describe "Promotion Adjustments" do
 
     it "ceasing to be eligible for a promotion with item total rule then becoming eligible again" do
       fill_in "Name", :with => "Spend over $50 and save $5"
-      select "Order contents changed", :from => "Event"
+      select2 "Order contents changed", :from => "Event Name"
       click_button "Create"
       page.should have_content("Editing Promotion")
 
@@ -388,7 +394,7 @@ describe "Promotion Adjustments" do
 
     it "only counting the most valuable promotion adjustment in an order" do
       fill_in "Name", :with => "$5 off"
-      select "Order contents changed", :from => "Event"
+      select2 "Order contents changed", :from => "Event Name"
       click_button "Create"
       page.should have_content("Editing Promotion")
       select "Create adjustment", :from => "Add action of type"
@@ -401,7 +407,7 @@ describe "Promotion Adjustments" do
       visit spree.admin_promotions_path
       click_link "New Promotion"
       fill_in "Name", :with => "10% off"
-      select "Order contents changed", :from => "Event"
+      select2 "Order contents changed", :from => "Event Name"
       click_button "Create"
       page.should have_content("Editing Promotion")
       select "Create adjustment", :from => "Add action of type"
