+++ promo/spec/spec_helper.rb
@@ -12,6 +12,7 @@ require 'ffaker'
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
 
 require 'spree/core/testing_support/factories'
+require 'spree/core/testing_support/capybara_ext'
 require 'spree/core/testing_support/authorization_helpers'
 
 require 'factories'
