+++ promo/spec/support/promotion_creation.rb
@@ -41,7 +41,7 @@ module PromotionCreation
     promotion_name = "Order's total > $#{order_min}, Discount #{order_discount}"
     fill_in "Name", :with => promotion_name
     fill_in "Usage Limit", :with => "100"
-    select "Coupon code added", :from => "Event"
+    select2 "Coupon code added", :from => "Event Name"
     fill_in "Code", :with => coupon_code
     click_button "Create"
     page.should have_content("Editing Promotion")
